GUARD-3571 Rename OrdersApiClient to OrdersApiService
